requirements.txt
absl-py==2.0.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
astunparse==1.6.3
asyncpg==0.30.0
attrs==23.1.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
boto3==1.34.159
botocore==1.34.159
build==1.2.2.post1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
crashtest==0.4.1
cryptography==42.0.8
ctransformers==0.2.27
databases==0.9.0
dataclasses-json==0.6.4
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-crontab==0.7.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dm-tree==0.1.8
dnspython==2.7.0
dulwich==0.22.8
easygui==0.98.3
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.115.8
fastjsonschema==2.21.1
filelock==3.13.1
findpython==0.6.3
firebase-admin==6.5.0
Flask==3.0.2
flatbuffers==23.5.26
form==0.0.1
frozendict==2.3.10
frozenlist==1.4.1
fsspec==2025.3.0
fst-pso==1.8.1
FuzzyTM==2.0.5
gast==0.5.4
get-papers-list-pubmed==0.1.0
get-papers-list-sakshi==0.1.0
get-papers-list-sakshii==0.1.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==23.0.0
h11==0.14.0
h5py==3.10.0
haversine==2.8.1
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.7
imageio==2.35.0
importlib-metadata==6.11.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==2.15.0
keyring==25.6.0
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langsmith==0.1.10
lazy_loader==0.4
libclang==16.0.6
llvmlite==0.43.0
load-dotenv==0.1.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.0
mdurl==0.1.2
miniful==0.0.6
ml-dtypes==0.2.0
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.7
networkx==3.3
numba==0.60.0
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.19.0
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.9.15
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pbs-installer==2025.3.11
peewee==3.17.0
pillow==10.3.0
pipenv==2023.12.1
pkginfo==1.12.1.2
platformdirs==4.1.0
pluggy==1.5.0
poetry==2.1.1
poetry-core==2.1.1
pooch==1.8.2
proto-plus==1.24.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pyFUME==0.2.25
Pygments==2.17.2
PyJWT==2.8.0
PyMatting==1.1.12
pyparsing==3.1.2
pypdf==5.4.0
pyproject_hooks==1.2.0
pyreadline3==3.4.1
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2023.3.post1
pywin32-ctypes==0.2.3
PyYAML==6.0.1
RapidFuzz==3.12.2
referencing==0.32.0
regex==2024.11.6
rembg==2.0.58
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
s3transfer==0.10.2
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.14.0
sentence-transformers==3.4.1
sentry-sdk==2.13.0
shellingham==1.5.4
simpful==2.11.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.5.0
starlette==0.45.3
streamlit==1.29.0
sympy==1.13.1
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.6.0
tifffile==2024.8.10
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.0
torch==2.6.0
tornado==6.4
tqdm==4.66.5
transformers==4.50.0
trove-classifiers==2025.3.13.13
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.0
validators==0.22.0
virtualenv==20.29.3
watchdog==3.0.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
yfinance==0.2.33
zipp==3.17.0
zstandard==0.23.0
